+++ cmake/platforms/gcc-linux-x64-AVX2.cmake
