+++ cmake/platforms/gcc-windows-x64-AVX2.cmake
