+++ cmake/platforms/gcc-windows-x64.cmake
