+++ cmake/platforms/gcc-x64-autodetect.cmake
