+++ src/platform/windows/win_memory.c
@@ -1,7 +1,9 @@
 #ifdef _WIN32
 
 #include "anoptic_memalign.h"
-#include <Windows.h>
+#include <malloc.h>
+
+// TODO: Figure out what this is lol
 
 // Windows-specific implementation of aligned_malloc as defined in the anoptic_memory API.
 void* anoptic_aligned_malloc(size_t size, size_t alignment) {
