+++ src/vulkan_backend/pipeline.h
@@ -25,17 +25,19 @@
 
 #ifndef STRUCTS_H
 #define STRUCTS_H
-#include "graphics/structs.h"
+#include "vulkan_backend/structs.h"
 #endif
 
-// Function interfaces
+// Pipeline-specific structs
+struct Buffer 
+{
+    uint32_t size;
+    char* data;
+};
 
-// Initializes Vulkan, returns a pointer to VulkanComponents, or NULL on failure
-VulkanComponents* initVulkan(GLFWwindow* window, VulkanComponents* components); 
 
-// A celebration
-void unInitVulkan();
+// Creates a render pass
+bool createRenderPass(VkDevice device, VkFormat swapChainImageFormat, VkRenderPass* renderPass);
 
-// Draws a single frame
-
-void drawFrame(VulkanComponents* components, GLFWwindow* window);
+// Creates a graphics pipeline
+VkPipeline createGraphicsPipeline(VkDevice device, VkExtent2D swapChainExtent, VkPipelineLayout *pipelineLayout, VkRenderPass renderPass);
