+++ src/vulkan_backend/structs.h
@@ -100,11 +100,11 @@ struct QueueFamilyIndices // Stores whether different queue families exist, and 
 
 typedef struct DeviceCapabilities // Add queue families, device extensions etc as they're implemented into compute tasks and render functions
 {
-		bool graphics;
-		bool compute;
-		bool transfer;
-		bool float64;
-		bool int64;
+	bool graphics;
+	bool compute;
+	bool transfer;
+	bool float64;
+	bool int64;
 } DeviceCapabilities;
 
 struct SwapChainSupportDetails
