+++ src/vulkan_backend/vulkanMaster.c
@@ -35,17 +35,17 @@
 
 #ifndef INSTANCE_H
 #define INSTANCE_H
-#include "graphics/instanceInit.h"
+#include "vulkan_backend/instanceInit.h"
 #endif
 
 #ifndef STRUCTS_H
 #define STRUCTS_H
-#include "graphics/structs.h"
+#include "vulkan_backend/structs.h"
 #endif
 
 #ifndef PIPELINE_H
 #define PIPELINE_H
-#include "graphics/pipeline.h"
+#include "vulkan_backend/pipeline.h"
 #endif
